+++ src/utility.h
@@ -93,7 +93,7 @@ public:
 
 bool is_keypad(int sc);
 
-void rotate_angle(float& angle, float shift);
+void rotate_angle(double& angle, double shift);
 
 double get_time();
